remove the deprecations so this PR is just the fixture changes and add the 6th deprecation fixture
