+++ packages/e2e-tests/fixtures/blocks/core__gallery.json
@@ -44,7 +44,7 @@
                     "isListItem": true
                 },
                 "innerBlocks": [],
-                "originalContent": "<li class=\"list-image\">\n\t\t\t<figure class=\"wp-block-image size-large\">\n\t\t\t\t<img\n\t\t\t\tsrc=\"http://wptest.local/wp-content/uploads/2020/09/test-image-edited-1024x682.jpg\"\"\n\t\t\t\talt=\"Image gallery image\" class=\"wp-image-704\"/>\n\t\t\t</figure>\n\t\t</li>"
+                "originalContent": "<li class=\"list-image\">\n\t\t\t<figure class=\"wp-block-image size-large\">\n\t\t\t\t<img\n\t\t\t\tsrc=\"http://wptest.local/wp-content/uploads/2020/09/test-image-edited-1024x682.jpg\"\n\t\t\t\talt=\"Image gallery image\" class=\"wp-image-704\"/>\n\t\t\t</figure>\n\t\t</li>"
             }
         ],
         "originalContent": "<figure class=\"wp-block-gallery columns-2 is-cropped\">\n\t<ul class=\"blocks-gallery-grid\">\n\t\t\n\t\t\n\t</ul>\n</figure>"
