+++ packages/e2e-tests/fixtures/blocks/core__gallery.parsed.json
@@ -29,9 +29,9 @@
                     "isListItem": true
                 },
                 "innerBlocks": [],
-                "innerHTML": "\n\t\t<li class=\"list-image\">\n\t\t\t<figure class=\"wp-block-image size-large\">\n\t\t\t\t<img\n\t\t\t\tsrc=\"http://wptest.local/wp-content/uploads/2020/09/test-image-edited-1024x682.jpg\"\"\n\t\t\t\talt=\"Image gallery image\" class=\"wp-image-704\"/>\n\t\t\t</figure>\n\t\t</li>\n\t\t",
+                "innerHTML": "\n\t\t<li class=\"list-image\">\n\t\t\t<figure class=\"wp-block-image size-large\">\n\t\t\t\t<img\n\t\t\t\tsrc=\"http://wptest.local/wp-content/uploads/2020/09/test-image-edited-1024x682.jpg\"\n\t\t\t\talt=\"Image gallery image\" class=\"wp-image-704\"/>\n\t\t\t</figure>\n\t\t</li>\n\t\t",
                 "innerContent": [
-                    "\n\t\t<li class=\"list-image\">\n\t\t\t<figure class=\"wp-block-image size-large\">\n\t\t\t\t<img\n\t\t\t\tsrc=\"http://wptest.local/wp-content/uploads/2020/09/test-image-edited-1024x682.jpg\"\"\n\t\t\t\talt=\"Image gallery image\" class=\"wp-image-704\"/>\n\t\t\t</figure>\n\t\t</li>\n\t\t"
+                    "\n\t\t<li class=\"list-image\">\n\t\t\t<figure class=\"wp-block-image size-large\">\n\t\t\t\t<img\n\t\t\t\tsrc=\"http://wptest.local/wp-content/uploads/2020/09/test-image-edited-1024x682.jpg\"\n\t\t\t\talt=\"Image gallery image\" class=\"wp-image-704\"/>\n\t\t\t</figure>\n\t\t</li>\n\t\t"
                 ]
             }
         ],
