+++ packages/e2e-tests/fixtures/blocks/core__gallery__deprecated-4.json
@@ -4,57 +4,37 @@
         "name": "core/gallery",
         "isValid": true,
         "attributes": {
-            "caption": "",
-            "imageCrop": true,
-            "sizeSlug": "large",
-            "imageCount": 3,
-            "allowResize": false,
-            "isListItem": true
-        },
-        "innerBlocks": [
-            {
-                "clientId": "_clientId_0",
-                "name": "core/image",
-                "isValid": true,
-                "attributes": {
+            "images": [
+                {
                     "url": "https://sergioestevaofolio.files.wordpress.com/2016/09/cropped-img_9054-1.jpg?w=190",
                     "alt": "",
-                    "caption": "",
-                    "id": 1421,
-                    "sizeSlug": "large",
-                    "isListItem": false
+                    "id": "1421",
+                    "caption": ""
                 },
-                "innerBlocks": []
-            },
-            {
-                "clientId": "_clientId_1",
-                "name": "core/image",
-                "isValid": true,
-                "attributes": {
+                {
                     "url": "https://sergioestevaofolio.files.wordpress.com/2017/09/cropped-l1001498-1.jpg?w=580",
                     "alt": "",
-                    "caption": "",
-                    "id": 1440,
-                    "sizeSlug": "large",
-                    "isListItem": false
+                    "id": "1440",
+                    "caption": ""
                 },
-                "innerBlocks": []
-            },
-            {
-                "clientId": "_clientId_2",
-                "name": "core/image",
-                "isValid": true,
-                "attributes": {
+                {
                     "url": "https://sergioestevaofolio.files.wordpress.com/2017/05/cropped-l1005945-2-2.jpg?w=580",
                     "alt": "",
-                    "caption": "",
-                    "id": 1362,
-                    "sizeSlug": "large",
-                    "isListItem": false
-                },
-                "innerBlocks": []
-            }
-        ],
+                    "id": "1362",
+                    "caption": ""
+                }
+            ],
+            "ids": [
+                1421,
+                1440,
+                1362
+            ],
+            "caption": "",
+            "imageCrop": true,
+            "linkTo": "none",
+            "align": "wide"
+        },
+        "innerBlocks": [],
         "originalContent": "<figure class=\"wp-block-gallery columns-3 is-cropped alignwide\">\n\t<ul class=\"blocks-gallery-grid\">\n\t\t<li class=\"blocks-gallery-item\">\n\t\t\t<figure>\n\t\t\t\t<img src=\"https://sergioestevaofolio.files.wordpress.com/2016/09/cropped-img_9054-1.jpg?w=190\" alt=\"\" data-id=\"1421\" class=\"wp-image-1421\"/>\n\t\t\t</figure>\n\t\t</li>\n\t\t<li class=\"blocks-gallery-item\">\n\t\t\t<figure>\n\t\t\t\t<img src=\"https://sergioestevaofolio.files.wordpress.com/2017/09/cropped-l1001498-1.jpg?w=580\" alt=\"\" data-id=\"1440\" class=\"wp-image-1440\"/>\n\t\t\t</figure>\n\t\t</li>\n\t\t<li class=\"blocks-gallery-item\">\n\t\t\t<figure>\n\t\t\t\t<img src=\"https://sergioestevaofolio.files.wordpress.com/2017/05/cropped-l1005945-2-2.jpg?w=580\" alt=\"\" data-id=\"1362\" class=\"wp-image-1362\"/>\n\t\t\t</figure>\n\t\t</li>\n\t</ul>\n</figure>"
     }
 ]
