+++ packages/e2e-tests/fixtures/blocks/core__gallery__deprecated-4.serialized.html
@@ -1,13 +1,3 @@
-<!-- wp:gallery {"imageCount":3} -->
-<figure class="wp-block-gallery columns-3 is-cropped"><ul class="blocks-gallery-grid"><!-- wp:image {"id":1421,"sizeSlug":"large"} -->
-<figure class="wp-block-image size-large"><img src="https://sergioestevaofolio.files.wordpress.com/2016/09/cropped-img_9054-1.jpg?w=190" alt="" class="wp-image-1421"/></figure>
-<!-- /wp:image -->
-
-<!-- wp:image {"id":1440,"sizeSlug":"large"} -->
-<figure class="wp-block-image size-large"><img src="https://sergioestevaofolio.files.wordpress.com/2017/09/cropped-l1001498-1.jpg?w=580" alt="" class="wp-image-1440"/></figure>
-<!-- /wp:image -->
-
-<!-- wp:image {"id":1362,"sizeSlug":"large"} -->
-<figure class="wp-block-image size-large"><img src="https://sergioestevaofolio.files.wordpress.com/2017/05/cropped-l1005945-2-2.jpg?w=580" alt="" class="wp-image-1362"/></figure>
-<!-- /wp:image --></ul></figure>
+<!-- wp:gallery {"linkTo":"none","align":"wide"} -->
+<figure class="wp-block-gallery alignwide columns-3 is-cropped"><ul class="blocks-gallery-grid"></ul></figure>
 <!-- /wp:gallery -->
